Major Changes

+++ gamemode/modules/doors/libraries/server.lua
@@ -78,9 +78,7 @@ function MODULE:SaveData()
         end
     end
 
-    lia.db.delete("doors", condition):next(function()
-        if #rows > 0 then return lia.db.bulkInsert("doors", rows) end
-    end)
+    lia.db.delete("doors", condition):next(function() if #rows > 0 then return lia.db.bulkInsert("doors", rows) end end)
 end
 
 function MODULE:InitPostEntity()
@@ -232,7 +230,7 @@ function MODULE:ToggleLock(client, door, state)
         end
 
         door:setLocked(state)
-    elseif (door:GetCreator() == client or client:hasPrivilege(L("Manage Doors")) or client:isStaffOnDuty()) and (door:IsVehicle() or door:isSimfphysCar()) then
+    elseif (door:GetCreator() == client or client:hasPrivilege(L("manageDoors")) or client:isStaffOnDuty()) and (door:IsVehicle() or door:isSimfphysCar()) then
         if state then
             door:Fire("lock")
             client:EmitSound("doors/door_latch3.wav")
